requirements.txt
google-generativeai
SpeechRecognition
pyttsx3
pyaudio
tqdm
streamlit
streamlit-mic-recorder
gtts
python-dotenv
markdown